backend/requirements.txt
streamlit==1.38.0
langchain==0.3.0
langchain-openai==0.2.0
pandas==2.2.2
plotly==5.22.0
pdfminer.six==20231228
httpx==0.27.0
reportlab==4.2.2
